+++ ft_printf/print_integer/putnbr_base.c
@@ -1,16 +1,16 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   nbr_utils.c                                        :+:      :+:    :+:   */
+/*   putnbr_base.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/07 18:48:21 by reasuke           #+#    #+#             */
-/*   Updated: 2024/02/23 22:02:33 by reasuke          ###   ########.fr       */
+/*   Updated: 2024/02/23 22:23:01 by reasuke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "ft_printf.h"
+#include "../ft_printf.h"
 
 static void	_putnbr_base_rec(uintmax_t uim_nb, size_t radix, const char *base)
 {
