+++ ft_printf/print_string/print_padding.c
@@ -1,0 +1,19 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   print_string copy 2.c                              :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/09/09 17:02:59 by reasuke           #+#    #+#             */
+/*   Updated: 2024/02/23 22:18:36 by reasuke          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../ft_printf.h"
+
+void	fpf_print_padding(char c, int len)
+{
+	while (len-- > 0)
+		ft_putchar_fd(c, STDOUT_FILENO);
+}
